apply jike topic optimize to topicSquare, close #742
